pyproject.toml
[project]
name = "anti_loss_tag"
version = "1.8.0"
description = "KT6368A BLE 防丢标签 Home Assistant 集成"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "MMMM", email = "noreply@example.com"}
]
keywords = ["home-assistant", "ble", "bluetooth", "anti-loss", "tag", "kt6368a"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
]
dependencies = [
    "bleak>=0.21.0",
    "bleak-retry-connector>=3.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-homeassistant-custom-component>=0.13.0",
]

[project.urls]
Homepage = "https://github.com/xxoolm/anti_loss_tag"
Documentation = "https://github.com/xxoolm/anti_loss_tag/blob/main/README.md"
Repository = "https://github.com/xxoolm/anti_loss_tag"
"Bug Tracker" = "https://github.com/xxoolm/anti_loss_tag/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

addopts = "-v --strict-markers --tb=short"
markers = [
    "asyncio: mark test as async test",
    "integration: mark test as integration test (slow)",
    "unit: mark test as unit test (fast)",
]

[tool.coverage.run]
source = ["custom_components/anti_loss_tag"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/archived/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
